Add battery saving features, slow heartbeat on lamp off

+++ src/colors/colors.h
@@ -223,7 +223,7 @@ class GenerateRainbowIndex : public IndexedColor
 {
     public:
     GenerateRainbowIndex(const uint8_t colorDivisions)
-    : _increment(UINT8_MAX / float(colorDivisions)), _currentPixelHue(0)
+    : _increment(UINT16_MAX / float(colorDivisions)), _currentPixelHue(0)
     {}
 
     uint32_t get_color(const uint16_t index, const uint16_t maxIndex) const override;
